StronglyTypedInterviewEvaluator: changed type from int to long in UpdateIntAnswer method KP-3738
--HG--
branch : csharp-business-language


Former-commit-id: ff1c36e3233fb246df0b15f11cdee4fce8f2fc14

+++ src/Core/SharedKernels/DataCollection/ExpressionProcessing/InterviewExpressionMate.cs
@@ -217,8 +217,8 @@ namespace WB.Core.SharedKernels.ExpressionProcessing
                 return invalidIdentities;
             }
         }
-        
-        public void UpdateIntAnswer(Guid questionId, decimal[] rosterVector, int answer)
+
+        public void UpdateIntAnswer(Guid questionId, decimal[] rosterVector, long answer)
         {
             throw new NotImplementedException();
         }
